+++ CHANGELOG.md
@@ -10,7 +10,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
-## [Unreleased]
+## [0.15.2] - 2024-07-24
 
 ### Fixed
 
@@ -388,6 +388,7 @@ guide][0.11.0 migration guide] for detailed instructions.
 [0.14.0]: https://github.com/dogmatiq/testkit/releases/tag/v0.14.0
 [0.15.0]: https://github.com/dogmatiq/testkit/releases/tag/v0.15.0
 [0.15.1]: https://github.com/dogmatiq/testkit/releases/tag/v0.15.1
+[0.15.2]: https://github.com/dogmatiq/testkit/releases/tag/v0.15.2
 
 <!-- version template
 ## [0.0.1] - YYYY-MM-DD